+++ src/utils/restaurantOwnershipValidator.ts
